adding create_implied_field() utility
