+++ fiftyone/core/odm/__init__.py
@@ -38,7 +38,6 @@ from .database import (
     bulk_write,
 )
 from .dataset import (
-    create_field,
     SampleFieldDocument,
     KeypointSkeleton,
     DatasetAppConfig,
@@ -67,6 +66,8 @@ from .utils import (
     serialize_value,
     deserialize_value,
     validate_field_name,
+    create_field,
+    create_implied_field,
     get_field_kwargs,
     get_implied_field_kwargs,
     validate_fields_match,
