+++ fiftyone/core/odm/mixins.py
@@ -17,10 +17,12 @@ from .database import get_db_conn
 from .dataset import create_field, SampleFieldDocument
 from .document import Document
 from .utils import (
-    serialize_value,
     deserialize_value,
-    validate_field_name,
+    serialize_value,
+    create_field,
+    create_implied_field,
     get_implied_field_kwargs,
+    validate_field_name,
     validate_fields_match,
 )
 
@@ -366,7 +368,7 @@ class DatasetMixin(object):
             ValueError: if a field in the schema is not compliant with an
                 existing field of the same name
         """
-        field = cls._create_implied_field(path, value, dynamic)
+        field = create_implied_field(path, value, dynamic=dynamic)
 
         return cls.merge_field_schema(
             {path: field},
@@ -398,12 +400,6 @@ class DatasetMixin(object):
             info=info,
             **kwargs,
         )
-
-    @classmethod
-    def _create_implied_field(cls, path, value, dynamic):
-        field_name = path.rsplit(".", 1)[-1]
-        kwargs = get_implied_field_kwargs(value, dynamic=dynamic)
-        return create_field(field_name, **kwargs)
 
     @classmethod
     def _rename_fields(cls, sample_collection, paths, new_paths):
